requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.2
django-debug-toolbar==4.3.0
django-filter==23.5
django-querycount==0.8.3
djangorestframework==3.14.0
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1